# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Domain:
    properties:
      ICP_reg:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      vendor:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      domains:
        items:
          $ref: '#/definitions/models.Domain'
        type: array
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      stuId:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.UserRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      Admin: can submit new domain, access and delete all domains, promte/demote user
        to Contributor, grant/deny all domains access to other Normal user
      Contributor: can submit new domain, access and delete own domain, grant/deny
        own domain access to other Normal user
      Normal: only can access granted domains
      SysAdmin: same as Admin, promte/demote user to Admin
    x-enum-varnames:
    - Normal
    - Contributor
    - Admin
    - SysAdmin
...
  web.Domain:
    properties:
      data: {}
      errors:
        type: string
      status:
        type: integer
    type: object
  web.DomainInfoUpdate:
    properties:
      ICP_reg:
        type: boolean
      api_id:
        type: string
      api_secret:
        type: string
      name:
        type: string
      vendor:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: makiras.x@outlook.com
    name: domain0
  description: Domain0 API
  license:
    name: MPL(mozilla public license)-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
  title: Domain0 API
  version: 0.0.1
paths:
  /api/v1/domain:
    get:
      description: |-
        List domains
        user can list all domains if user role level is admin
        user can list domains which user has read access if user role level is not admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Domain'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
      summary: List domains
      tags:
      - domain
    post:
      consumes:
      - application/json
      description: |-
        Create domain
        user must have contributor role or higher
      parameters:
      - description: domain info
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/web.DomainInfoUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/web.DomainInfoUpdate'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
      summary: Create domain
      tags:
      - domain
  /api/v1/domain/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete domain
        user must have owner role to domain or be admin
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete domain
      tags:
      - domain
    get:
      description: |-
        Get domain by id
        user must have read permission to domain or be admin
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Get domain by id
      tags:
      - domain
    put:
      consumes:
      - application/json
      description: |-
        Update domain
        user must have manager role to domain or be admin
        **ICP_reg param can't be updated**
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: domain info
        in: body
        name: domain
        required: true
        schema:
          $ref: '#/definitions/web.DomainInfoUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/web.DomainInfoUpdate'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: string
              type: object
      summary: Update domain
      tags:
      - domain
...
schemes:
- http
- https
swagger: "2.0"